requirements.txt
# Python Dependencies for Trading Platform
# Phase 2: Data Collection

# Core
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
PyYAML==6.0.1

# Async & Concurrency
asyncio-throttle==1.0.2

# Kafka
aiokafka==0.10.0
kafka-python==2.3.0

# Databases
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1

# S3/MinIO
boto3==1.34.10
botocore==1.34.10

# Data Collection - RSS
feedparser==6.0.10

# Data Collection - Reddit
praw==7.7.1

# Data Collection - Market Data
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2
ta-lib==0.4.28

# HTTP & Web
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
hypothesis==6.92.1

# Development
black==23.12.1
flake8==7.0.0
mypy==1.7.1
isort==5.13.2

# NLP
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
langdetect==1.0.9

# AI Providers (Phase 3)
langchain==0.1.0
langgraph==0.0.20
anthropic==0.75.0
openai==1.6.1

# LangSmith (LLM Tracing & Monitoring)
langsmith==0.5.2
langchain-anthropic==1.3.0
langchain-core==1.2.5

# Market Data (for future phases)
# yfinance==0.2.33
# ccxt==4.1.92
# ib-insync==0.9.86

# Analysis (for future phases)
# pandas==2.1.4
# numpy==1.26.2
# ta-lib==0.4.28

# API (for future phases)
# fastapi==0.108.0
# uvicorn==0.25.0
# websockets==12.0
